DevListComponent desing completed
The overall desing for the devList is complete. It looks a little bland as-is but its ok and intuitive for now.

I might add the RC logo somewhere to make the page more dynamic and make the app more consistent (logo in every page).

Noticed bug : I made the CSS so that the cursor is in 'dragging' state when we drag an item. It works fine on the very start of the dragging event, but then when actually dragging an item, the cursor doesn't stay on its 'dragging' state.