Improved tiny screens display
Tweaked a couple of things in the RandomChooserComponent so that the image is more reactive and all the buttons are still visible on really tiny screens.

NOTE : 
Images that aren't perfect squares won't appear as perfect circles in the random chooser component, so we will need something to crop the pictures to a square eventually.